+++ server/src/main/java/com/continuuity/loom/scheduler/task/ClusterService.java
@@ -15,6 +15,7 @@
  */
 package com.continuuity.loom.scheduler.task;
 
+import com.continuuity.loom.admin.ClusterTemplate;
 import com.continuuity.loom.cluster.Cluster;
 import com.continuuity.loom.cluster.Node;
 import com.continuuity.loom.codec.json.JsonSerde;
@@ -23,12 +24,14 @@ import com.continuuity.loom.common.queue.TrackingQueue;
 import com.continuuity.loom.common.zookeeper.lib.ZKInterProcessReentrantLock;
 import com.continuuity.loom.conf.Constants;
 import com.continuuity.loom.http.AddServicesRequest;
-import com.continuuity.loom.layout.ClusterLayoutUpdater;
+import com.continuuity.loom.layout.ClusterLayout;
+import com.continuuity.loom.layout.InvalidClusterException;
 import com.continuuity.loom.layout.Solver;
 import com.continuuity.loom.management.LoomStats;
 import com.continuuity.loom.scheduler.ClusterAction;
 import com.continuuity.loom.scheduler.SolverRequest;
 import com.continuuity.loom.store.ClusterStore;
+import com.continuuity.loom.store.EntityStore;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableSet;
 import com.google.gson.Gson;
@@ -50,7 +53,8 @@ public class ClusterService {
   private static final Logger LOG = LoggerFactory.getLogger(ClusterService.class);
   private static final Gson GSON = new JsonSerde().getGson();
 
-  private final ClusterStore store;
+  private final ClusterStore clusterStore;
+  private final EntityStore entityStore;
   private final TrackingQueue clusterQueue;
   private final TrackingQueue solverQueue;
   private final ZKClient zkClient;
@@ -58,10 +62,15 @@ public class ClusterService {
   private final Solver solver;
 
   @Inject
-  private ClusterService(ClusterStore store, @Named(Constants.Queue.CLUSTER) TrackingQueue clusterQueue,
-                         @Named(Constants.Queue.SOLVER) TrackingQueue solverQueue, ZKClient zkClient,
-                         LoomStats loomStats, Solver solver) {
-    this.store = store;
+  public ClusterService(ClusterStore clusterStore,
+                        EntityStore entityStore,
+                        @Named(Constants.Queue.CLUSTER) TrackingQueue clusterQueue,
+                        @Named(Constants.Queue.SOLVER) TrackingQueue solverQueue,
+                        ZKClient zkClient,
+                        LoomStats loomStats,
+                        Solver solver) {
+    this.clusterStore = clusterStore;
+    this.entityStore = entityStore;
     this.clusterQueue = clusterQueue;
     this.solverQueue = solverQueue;
     this.zkClient = ZKClients.namespace(zkClient, Constants.LOCK_NAMESPACE);
@@ -81,15 +90,15 @@ public class ClusterService {
     lock.acquire();
     try {
       Cluster cluster = getUserCluster(clusterId, userId);
-      JobId deleteJobId = store.getNewJobId(clusterId);
+      JobId deleteJobId = clusterStore.getNewJobId(clusterId);
       ClusterJob deleteJob = new ClusterJob(deleteJobId, ClusterAction.CLUSTER_DELETE);
       deleteJob.setJobStatus(ClusterJob.Status.RUNNING);
       cluster.setLatestJobId(deleteJobId.getId());
       cluster.setStatus(Cluster.Status.PENDING);
 
       LOG.debug("Writing cluster {} to store with delete job {}", clusterId, deleteJobId);
-      store.writeCluster(cluster);
-      store.writeClusterJob(deleteJob);
+      clusterStore.writeCluster(cluster);
+      clusterStore.writeClusterJob(deleteJob);
 
       loomStats.getClusterStats().incrementStat(ClusterAction.CLUSTER_DELETE);
       clusterQueue.add(new Element(clusterId, ClusterAction.CLUSTER_DELETE.name()));
@@ -98,6 +107,17 @@ public class ClusterService {
     }
   }
 
+  /**
+   * Put in a request to reconfigure services on an active cluster.
+   * Throws a {@link MissingClusterException} if no cluster owned by the user is found and an
+   * {@link IllegalStateException} if the cluster is not in a state where the action can be performed.
+   *
+   * @param clusterId Id of cluster to reconfigure.
+   * @param userId User that is trying to reconfigure the cluster.
+   * @param restartServices Whether or not services should be restarted as part of the reconfigure.
+   * @param config New value of the config to use.
+   * @throws Exception
+   */
   public void requestClusterReconfigure(String clusterId, String userId, boolean restartServices, JsonObject config)
     throws Exception {
     ZKInterProcessReentrantLock lock = new ZKInterProcessReentrantLock(zkClient, "/" + clusterId);
@@ -110,7 +130,7 @@ public class ClusterService {
       if (!Cluster.Status.CONFIGURABLE_STATES.contains(cluster.getStatus())) {
         throw new IllegalStateException("cluster " + clusterId + " is not in a configurable state");
       }
-      JobId configureJobId = store.getNewJobId(clusterId);
+      JobId configureJobId = clusterStore.getNewJobId(clusterId);
 
       ClusterAction action =
         restartServices ? ClusterAction.CLUSTER_CONFIGURE_WITH_RESTART : ClusterAction.CLUSTER_CONFIGURE;
@@ -121,8 +141,8 @@ public class ClusterService {
 
       LOG.debug("Writing cluster {} to store with configure job {}", clusterId, configureJobId);
       cluster.setConfig(config);
-      store.writeCluster(cluster);
-      store.writeClusterJob(configureJob);
+      clusterStore.writeCluster(cluster);
+      clusterStore.writeClusterJob(configureJob);
 
       loomStats.getClusterStats().incrementStat(action);
       clusterQueue.add(new Element(clusterId, action.name()));
@@ -131,6 +151,17 @@ public class ClusterService {
     }
   }
 
+  /**
+   * Put in a request to perform an action on a specific service on an active cluster, or on all cluster services.
+   * Throws a {@link MissingClusterException} if no cluster owned by the user is found and an
+   * {@link IllegalStateException} if the cluster is not in a state where the action can be performed.
+   *
+   * @param clusterId Id of cluster perform the service action on.
+   * @param userId User that is trying to perform an action on the cluster service.
+   * @param action Action to perform on the service.
+   * @param service Service to perform the action on. Null means perform the action on all services.
+   * @throws Exception
+   */
   public void requestServiceRuntimeAction(String clusterId, String userId, ClusterAction action, String service)
     throws Exception {
     Preconditions.checkArgument(ClusterAction.SERVICE_RUNTIME_ACTIONS.contains(action),
@@ -146,14 +177,14 @@ public class ClusterService {
         throw new IllegalStateException(
           "cluster " + clusterId + " is not in a state where service actions can be performed");
       }
-      JobId jobId = store.getNewJobId(clusterId);
+      JobId jobId = clusterStore.getNewJobId(clusterId);
 
       ClusterJob job = new ClusterJob(jobId, action, service == null ? null : ImmutableSet.of(service), null);
       job.setJobStatus(ClusterJob.Status.RUNNING);
       cluster.setLatestJobId(job.getJobId());
       cluster.setStatus(Cluster.Status.PENDING);
-      store.writeCluster(cluster);
-      store.writeClusterJob(job);
+      clusterStore.writeCluster(cluster);
+      clusterStore.writeClusterJob(job);
 
       loomStats.getClusterStats().incrementStat(action);
       clusterQueue.add(new Element(clusterId, action.name()));
@@ -162,6 +193,16 @@ public class ClusterService {
     }
   }
 
+  /**
+   * Put in a request to add services to an active cluster.
+   * Throws a {@link MissingClusterException} if no cluster owned by the user is found and an
+   * {@link IllegalStateException} if the cluster is not in a state where the action can be performed.
+   *
+   * @param clusterId Id of cluster to add services to.
+   * @param userId User that is trying to add services to the cluster.
+   * @param addRequest Request to add services.
+   * @throws Exception
+   */
   public void requestAddServices(String clusterId, String userId, AddServicesRequest addRequest)
     throws Exception {
     ZKInterProcessReentrantLock lock = new ZKInterProcessReentrantLock(zkClient, "/" + clusterId);
@@ -175,7 +216,7 @@ public class ClusterService {
         throw new IllegalStateException(
           "cluster " + clusterId + " is not in a state where services can be added to it.");
       }
-      JobId jobId = store.getNewJobId(clusterId);
+      JobId jobId = clusterStore.getNewJobId(clusterId);
       solver.validateServicesToAdd(cluster, addRequest.getServices());
 
       ClusterAction action = ClusterAction.ADD_SERVICES;
@@ -183,12 +224,58 @@ public class ClusterService {
       job.setJobStatus(ClusterJob.Status.RUNNING);
       cluster.setLatestJobId(job.getJobId());
       cluster.setStatus(Cluster.Status.PENDING);
-      store.writeCluster(cluster);
-      store.writeClusterJob(job);
+      clusterStore.writeCluster(cluster);
+      clusterStore.writeClusterJob(job);
 
       loomStats.getClusterStats().incrementStat(action);
       SolverRequest solverRequest = new SolverRequest(SolverRequest.Type.ADD_SERVICES, GSON.toJson(addRequest));
       solverQueue.add(new Element(clusterId, GSON.toJson(solverRequest)));
+    } finally {
+      lock.release();
+    }
+  }
+
+  /**
+   * Sync the template of the cluster to the current version of the template.
+   * Throws a {@link MissingClusterException} if no cluster owned by the user is found, an
+   * {@link IllegalStateException} if the cluster is not in a state where it can by synced, or a
+   * {@link MissingEntityException} if the template no longer exists.
+   *
+   * @param clusterId Id of the cluster to sync.
+   * @param userId Id of the user trying to sync the cluster.
+   */
+  public void syncClusterToCurrentTemplate(String clusterId, String userId) throws Exception {
+    ZKInterProcessReentrantLock lock = new ZKInterProcessReentrantLock(zkClient, "/" + clusterId);
+    lock.acquire();
+    try {
+      Cluster cluster = getUserCluster(clusterId, userId);
+      if (cluster == null) {
+        throw new MissingClusterException("cluster " + clusterId + " owned by user " + userId + " does not exist");
+      }
+      if (cluster.getStatus() != Cluster.Status.ACTIVE) {
+        throw new IllegalStateException(
+          "cluster " + clusterId + " is not in a state where services can be added to it.");
+      }
+
+      String templateName = cluster.getClusterTemplate().getName();
+      ClusterTemplate currentTemplate = entityStore.getClusterTemplate(templateName);
+      if (currentTemplate == null) {
+        LOG.info("tried to sync template {} for cluster {}, but the template no longer exists",
+                 templateName, clusterId);
+        throw new MissingEntityException("template " + templateName + " no longer exists");
+      }
+      Set<Node> clusterNodes = clusterStore.getClusterNodes(clusterId, userId);
+      if (clusterNodes.isEmpty()) {
+        throw new MissingEntityException("could not find cluster nodes");
+      }
+      ClusterLayout clusterLayout = ClusterLayout.fromNodes(clusterNodes, currentTemplate.getConstraints());
+      // don't allow updating the template if it would result in an invalid cluster.
+      if (!clusterLayout.isCompatibleWithTemplate(currentTemplate)) {
+        throw new InvalidClusterException("updating the template would result in an invalid cluster");
+      }
+      // all good, update the template and save it
+      cluster.setClusterTemplate(currentTemplate);
+      clusterStore.writeCluster(cluster);
     } finally {
       lock.release();
     }
@@ -236,7 +323,7 @@ public class ClusterService {
 
       cluster.setExpireTime(expireTime);
       LOG.debug("Prolonging lease of cluster {} by {} to {}", clusterId, expireTime, cluster.getExpireTime());
-      store.writeCluster(cluster);
+      clusterStore.writeCluster(cluster);
     } finally {
       lock.release();
     }
@@ -253,9 +340,9 @@ public class ClusterService {
   public Cluster getUserCluster(String clusterId, String userId) throws Exception {
     Cluster cluster;
     if (userId.equals(Constants.ADMIN_USER) || userId.equals(Constants.SYSTEM_USER)) {
-      cluster = store.getCluster(clusterId);
+      cluster = clusterStore.getCluster(clusterId);
     } else {
-      cluster = store.getCluster(clusterId, userId);
+      cluster = clusterStore.getCluster(clusterId, userId);
     }
     return cluster;
   }
@@ -270,9 +357,9 @@ public class ClusterService {
    */
   public List<ClusterJob> getClusterJobs(String clusterId, String userId) throws Exception {
     if (userId.equals(Constants.ADMIN_USER) || userId.equals(Constants.SYSTEM_USER)) {
-      return store.getClusterJobs(clusterId, -1);
+      return clusterStore.getClusterJobs(clusterId, -1);
     } else {
-      return store.getClusterJobs(clusterId, userId, -1);
+      return clusterStore.getClusterJobs(clusterId, userId, -1);
     }
   }
 
@@ -286,9 +373,9 @@ public class ClusterService {
   public List<Cluster> getAllUserClusters(String userId) throws Exception {
     List<Cluster> clusters;
     if (userId.equals(Constants.ADMIN_USER) || userId.equals(Constants.SYSTEM_USER)) {
-      clusters = store.getAllClusters();
+      clusters = clusterStore.getAllClusters();
     } else {
-      clusters = store.getAllClusters(userId);
+      clusters = clusterStore.getAllClusters(userId);
     }
     return clusters;
   }
@@ -303,9 +390,9 @@ public class ClusterService {
    */
   public Set<Node> getClusterNodes(String clusterId, String userId) throws Exception {
     if (userId.equals(Constants.ADMIN_USER) || userId.equals(Constants.SYSTEM_USER)) {
-      return store.getClusterNodes(clusterId);
+      return clusterStore.getClusterNodes(clusterId);
     } else {
-      return store.getClusterNodes(clusterId, userId);
+      return clusterStore.getClusterNodes(clusterId, userId);
     }
   }
 }